Add-most_populars +

+++ src/main/java/ru/yandex/practicum/filmorate/storage/InMemoryFilmStorage.java
@@ -58,7 +58,7 @@ public class InMemoryFilmStorage implements FilmDAO {
     }
 
     @Override
-    public List<Film> getPopularFilm(Integer count, Integer genreId, Integer year) {
+    public List<Film> getPopularFilm(Integer count, Long genreId, Integer year) {
         return findAll()
                 .stream()
                 .sorted((film1, film2) -> film2.getLikes().size() - film1.getLikes().size())
